pyproject.toml
[tool.poetry]
name = "learnhub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-filter = "^24.2"
django = "^5.0.6"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^70.1.1"
django-cors-headers = "^4.4.0"
stripe = "^10.1.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"
gevent = "^24.2.1"
coverage = "^7.5.4"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
coverage = "^7.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
